Merging to Bleeding Edge

+++ apothecary/formulas/assimp.sh
@@ -7,20 +7,24 @@
 # uses CMake
 
 # define the version
-VER=5.2.5
+VER=5.2.10
 
 # tools for git use
-GIT_URL=
-# GIT_URL=https://github.com/assimp/assimp.git
+GIT_URL=https://github.com/danoli3/assimp
 GIT_TAG=
 
 FORMULA_TYPES=( "osx" "ios" "tvos" "android" "emscripten" "vs" )
 
+FORMULA_DEPENDS=( "zlib" )
+
 # download the source code and unpack it into LIB_NAME
 function download() {
 
-    # stable release from source forge
-    curl -LO "https://github.com/assimp/assimp/archive/v$VER.zip"
+    echo "Downloading Assimp $VER"
+    # stable release from GitHub
+    echo "From $GIT_URL/archive/refs/tags/v$VER.zip"
+    curl -LO "$GIT_URL/archive/refs/tags/v$VER.zip"
+
     unzip -oq "v$VER.zip"
     mv "assimp-$VER" assimp
     rm "v$VER.zip"
@@ -31,42 +35,32 @@ function download() {
 function prepare() {
     echo "Prepare"
 
-    # if [ "$TYPE" == "ios" ] ; then
-
-    #   # # patch outdated Makefile.osx provided with FreeImage, check if patch was applied first
-    #   # if patch -p0 -u -N --dry-run --silent < $FORMULA_DIR/assimp.ios.patch 2>/dev/null ; then
-    #   #   patch -p0 -u < $FORMULA_DIR/assimp.ios.patch
-    #   # fi
-
-    # fi
+    apothecaryDependencies download
+    
+    apothecaryDepend prepare zlib
+    apothecaryDepend build zlib
+    apothecaryDepend copy zlib
+   
 }
 
 # executed inside the lib src dir
 function build() {
-
+    LIBS_ROOT=$(realpath $LIBS_DIR)
     rm -f CMakeCache.txt || true
-
-    local IOS_ARCHS
-    if [[ "${TYPE}" == "tvos" ]]; then
-        IOS_ARCHS="x86_64 arm64"
-    elif [[ "$TYPE" == "ios" ]]; then
-        IOS_ARCHS="x86_64 armv7 arm64" #armv7s
-    fi
-
     if [[ "$TYPE" == "ios" || "$TYPE" == "tvos" ]] ; then
         if [[ "$TYPE" == "tvos" ]]; then
             export IOS_MIN_SDK_VER=9.0
         fi
-        echo "building $TYPE | $IOS_ARCHS"
+        echo "building $TYPE"
         cd ./port/iOS/
-        ./build.sh --stdlib=libc++ --archs="armv7 armv7s arm64 x86_64" IOS_SDK_VERSION=$IOS_MIN_SDK_VER
+        ./build.sh --stdlib=libc++ --std=c++17 --archs="armv7 arm64 x86_64" IOS_SDK_VERSION=$IOS_MIN_SDK_VER
         echo "--------------------"
 
         echo "Completed Assimp for $TYPE"
     fi
 
     if [ "$TYPE" == "osx" ] ; then
-
+        find ./ -name "*.o" -type f -delete
         # warning, assimp on github uses the ASSIMP_ prefix for CMake options ...
         # these may need to be updated for a new release
         local buildOpts="
@@ -85,39 +79,43 @@ function build() {
         make assimp -j${PARALLEL_MAKE}
 
     elif [ "$TYPE" == "vs" ] ; then
-
-        unset TMP
-        unset TEMP
-        #architecture selection inspired int he tess formula, shouldn't build both architectures in the same run?
-        echo "building $TYPE | $ARCH | $VS_VER"
+        find ./ -name "*.o" -type f -delete
+        #architecture selection inspired int he tess formula, shouldn't build both architectures in the same run
+        echo "building $TYPE | $ARCH | $VS_VER | vs: $VS_VER_GEN"
         echo "--------------------"
+        GENERATOR_NAME="Visual Studio ${VS_VER_GEN}"     
 
-        local buildOpts="
+        ZLIB_ROOT="$LIBS_ROOT/zlib/"
+        ZLIB_INCLUDE_DIR="$LIBS_ROOT/zlib/include"
+        ZLIB_LIBRARY="$LIBS_ROOT/zlib/lib/$TYPE/zlib.lib"   
+
+        mkdir -p "build_${TYPE}_${ARCH}"
+        cd "build_${TYPE}_${ARCH}"
+
+        DEFS="
+            -DCMAKE_BUILD_TYPE=Release \
+            -DCMAKE_C_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD_REQUIRED=ON \
+            -DCMAKE_CXX_EXTENSIONS=OFF 
             -DBUILD_SHARED_LIBS=OFF
             -DASSIMP_BUILD_TESTS=0
             -DASSIMP_BUILD_SAMPLES=0
-            -DASSIMP_BUILD_3MF_IMPORTER=0
-            -DLIBRARY_SUFFIX=${ARCH}"
-        local generatorName="Visual Studio "
-        generatorName+=$VS_VER
-        if [ $ARCH == 32 ] ; then
-            mkdir -p build_vs_32
-            cd build_vs_32
-            cmake .. -G "$generatorName" $buildOpts
-            vs-build "Assimp.sln" build "Release|Win32"
-        elif [ $ARCH == 64 ] ; then
-            mkdir -p build_vs_64
-            cd build_vs_64
-            if [ $VS_VER == 15 ] ; then
-                generatorName+=' Win64'
-                cmake .. -G "$generatorName" $buildOpts
-            else
-                cmake .. -G "$generatorName" -A x64 $buildOpts
-            fi
-            vs-build "Assimp.sln" build "Release|x64"
-            vs-build "Assimp.sln" build "Debug|x64"
-        fi
-        cd ..
+            -DASSIMP_BUILD_3MF_IMPORTER=0"
+
+        cmake .. ${DEFS} \
+            -A "${PLATFORM}" \
+            ${CMAKE_WIN_SDK} \
+            -G "${GENERATOR_NAME}" \
+            -DASSIMP_BUILD_ZLIB=OFF \
+            -DZLIB_ROOT=${ZLIB_ROOT} \
+            -DZLIB_INCLUDE_DIR=${ZLIB_INCLUDE_DIR} \
+            -DZLIB_LIBRARY=${ZLIB_LIBRARY}
+
+        cmake --build . --config Release
+
+        cd ..      
+       
         #cleanup to not fail if the other platform is called
         rm -f CMakeCache.txt
         echo "--------------------"
@@ -129,12 +127,16 @@ function build() {
 
     elif [ "$TYPE" == "android" ] ; then
 
-        source ../../android_configure.sh $ABI
+        ANDROID_API=24
+        ANDROID_PLATFORM=android-${ANDROID_API}
+
+        source ../../android_configure.sh $ABI cmake
+
 								
         #stuff to remove when we upgrade android
         #android complains about abs being ambigious - pfffft
-        sed -i -e 's/abs(/(int)fabs(/g' include/assimp/Hash.h
-        sed -i -e '/string_view/d' code/AssetLib/Obj/ObjFileParser.cpp
+        #sed -i -e 's/abs(/(int)fabs(/g' include/assimp/Hash.h
+        #sed -i -e '/string_view/d' code/AssetLib/Obj/ObjFileParser.cpp
 
         if [ "$ABI" == "armeabi-v7a" ]; then
             export HOST=armv7a-linux-android
@@ -184,20 +186,89 @@ function build() {
                 -DANDROID_STL=c++_static
                 -DANDROID_NATIVE_API_LEVEL=$ANDROID_PLATFORM
                 -DCMAKE_INSTALL_PREFIX=install"
+        elif [ "$ABI" == "x86_64" ]; then
+            export HOST=x86_64-linux-android
+            local buildOpts="
+                -DBUILD_SHARED_LIBS=OFF
+                -DASSIMP_BUILD_TESTS=0
+                -DASSIMP_BUILD_SAMPLES=0
+                -DASSIMP_BUILD_3MF_IMPORTER=0
+                -DASSIMP_BUILD_ZLIB=1
+                -DANDROID_NDK=$NDK_ROOT
+                -DCMAKE_TOOLCHAIN_FILE=$ANDROID_CMAKE_TOOLCHAIN
+                -DCMAKE_BUILD_TYPE=Release
+                -DANDROID_ABI=$ABI
+                -DANDROID_STL=c++_static
+                -DANDROID_NATIVE_API_LEVEL=$ANDROID_PLATFORM
+                -DCMAKE_INSTALL_PREFIX=install"
         fi
- # -- Enabled formats: AMF 3DS AC ASE ASSBIN ASSXML B3D BVH COLLADA DXF CSM HMP IRRMESH IRR LWO LWS MD2 MD3 MD5 MDC MDL NFF NDO OFF OBJ OGRE OPENGEX PLY MS3D COB BLEND IFC XGL FBX Q3D Q3BSP RAW SIB SMD TERRAGEN 3D X X3D GLTF 3MF MMD
+        
+        find ./ -name "*.o" -type f -delete
+        
+        mkdir -p "build_${TYPE}_${ABI}"
+        cd "build_${TYPE}_${ABI}"
 
-            # -DASSIMP_BUILD_STL_IMPORTER=0
-            # -DASSIMP_BUILD_BLEND_IMPORTER=0
-        rm -rf build_android
-        mkdir -p build_android
-        cd build_android
-        cmake -G 'Unix Makefiles' $buildOpts -DCMAKE_C_FLAGS="-O3 -DNDEBUG ${CFLAGS}" -DCMAKE_CXX_FLAGS="-O3 -DNDEBUG ${CFLAGS}" -DCMAKE_LD_FLAGS="$LDFLAGS" ..
-        make assimp -j${PARALLEL_MAKE}
+        rm -f CMakeCache.txt
+
+
+        export CFLAGS=""
+        export CPPFLAGS=""
+        export LDFLAGS=""
+        export CMAKE_LDFLAGS="$LDFLAGS"
+        
+        cmake -S .. -DCMAKE_TOOLCHAIN_FILE=${NDK_ROOT}/build/cmake/android.toolchain.cmake \
+            $buildOpts \
+            -DCMAKE_C_COMPILER=${CC} \
+            -DASSIMP_ANDROID_JNIIOSYSTEM=OFF \
+            -DANDROID_STL=c++_shared \
+            -DCMAKE_CXX_COMPILER_RANLIB=${RANLIB} \
+            -DCMAKE_C_COMPILER_RANLIB=${RANLIB} \
+            -DCMAKE_CXX_COMPILER_AR=${AR} \
+            -DCMAKE_C_COMPILER_AR=${AR} \
+            -DAI_CONFIG_ANDROID_JNI_ASSIMP_MANAGER_SUPPORT=OFF \
+            -DCMAKE_CXX_FLAGS="-fvisibility-inlines-hidden -O3 -fPIC -Wno-implicit-function-declaration" \
+            -DCMAKE_C_FLAGS="-fvisibility-inlines-hidden -O3 -fPIC -Wno-implicit-function-declaration" \
+            -DCMAKE_EXE_LINKER_FLAGS=" -Wl,--hash-style=both" \
+            -DCMAKE_SHARED_LINKER_FLAGS="-Wl,--hash-style=both" \
+            -DCMAKE_MODULE_LINKER_FLAGS=" -Wl,--hash-style=both" \
+            -DCMAKE_CXX_STANDARD_LIBRARIES=${LIBS} \
+            -DCMAKE_C_STANDARD_LIBRARIES=${LIBS} \
+            -DCMAKE_STATIC_LINKER_FLAGS="${LDFLAGS} ${NDK_ROOT}/sources/cxx-stl/llvm-libc++/libs/${ABI}/libc++_shared.so ${NDK_ROOT}/sources/cxx-stl/llvm-libc++/libs/${ABI}/libc++abi.a " \
+            -DANDROID_NATIVE_API_LEVEL=${ANDROID_API} \
+            -DCMAKE_SYSROOT=$SYSROOT \
+            -DCMAKE_C_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD_REQUIRED=ON \
+            -DCMAKE_CXX_EXTENSIONS=OFF \
+            -DANDROID_TOOLCHAIN=clang++ \
+            -DBUILD_SHARED_LIBS=OFF \
+            -DASSIMP_BUILD_STATIC_LIB=1 \
+            -DASSIMP_BUILD_TESTS=0 \
+            -DASSIMP_BUILD_SAMPLES=0 \
+            -DASSIMP_BUILD_STL_IMPORTER=0 \
+            -DASSIMP_BUILD_BLEND_IMPORTER=0 \
+            -DASSIMP_BUILD_3MF_IMPORTER=0 \
+            -DASSIMP_ENABLE_BOOST_WORKAROUND=1 \
+            -DCMAKE_SYSROOT=$SYSROOT \
+            -DANDROID_NDK=$NDK_ROOT \
+            -DCMAKE_BUILD_TYPE=Release \
+            -DANDROID_ABI=$ABI \
+            -DANDROID_STL=c++_shared \
+            -DANDROID_PLATFORM=$ANDROID_PLATFORM \
+            -DANDROID_NATIVE_API_LEVEL=$ANDROID_PLATFORM \
+            -DCMAKE_INSTALL_PREFIX=install \
+            -DCMAKE_RUNTIME_OUTPUT_DIRECTORY="build_$ABI" \
+            -G 'Unix Makefiles' .
+        make clean
+        make -j${PARALLEL_MAKE} VERBOSE=1
         cd ..
-
+    
     elif [ "$TYPE" == "emscripten" ] ; then
+        find ./ -name "*.o" -type f -delete
 
+        ZLIB_ROOT="$LIBS_ROOT/zlib/"
+        ZLIB_INCLUDE_DIR="$LIBS_ROOT/zlib/include"
+        ZLIB_LIBRARY="$LIBS_ROOT/zlib/lib/$TYPE/zlib.a"
         # warning, assimp on github uses the ASSIMP_ prefix for CMake options ...
         # these may need to be updated for a new release
         local buildOpts="
@@ -205,11 +276,27 @@ function build() {
             -DASSIMP_BUILD_TESTS=0
             -DASSIMP_BUILD_SAMPLES=0
             -DASSIMP_BUILD_3MF_IMPORTER=0"
-        mkdir -p build_emscripten
-        cd build_emscripten
-        emcmake cmake -G 'Unix Makefiles' $buildOpts -DCMAKE_C_FLAGS="-O3 -DNDEBUG -pthread" -DCMAKE_CXX_FLAGS="-O3 -DNDEBUG -pthread" ..
-        emmake make assimp -j${PARALLEL_MAKE}
+        mkdir -p build_$TYPE
+        cd build_$TYPE
+        $EMSDK/upstream/emscripten/emcmake cmake .. \
+            -B . \
+            $buildOpts \
+            -DCMAKE_C_FLAGS="-O3 -DNDEBUG -DUSE_PTHREADS=1 -I${ZLIB_INCLUDE_DIR}" \
+            -DCMAKE_CXX_FLAGS="-O3 -DNDEBUG -DUSE_PTHREADS=1 -I${ZLIB_INCLUDE_DIR}" \
+            -DCMAKE_BUILD_TYPE=Release \
+            -DCMAKE_INCLUDE_OUTPUT_DIRECTORY=include \
+            -DCMAKE_INSTALL_INCLUDEDIR=include \
+            -DCMAKE_C_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD=17 \
+            -DCMAKE_CXX_STANDARD_REQUIRED=ON \
+            -DASSIMP_BUILD_ZLIB=ON \
+            -DZLIB_ROOT=${ZLIB_ROOT} \
+            -DZLIB_INCLUDE_DIR=${ZLIB_INCLUDE_DIR} \
+            -DZLIB_LIBRARIES=${ZLIB_LIBRARY} 
+
+        cmake --build . --config Release
         cd ..
+
     fi
 }
 
@@ -225,21 +312,10 @@ function copy() {
 
     # libs
     mkdir -p $1/lib/$TYPE
-    if [ "$TYPE" == "vs" ] ; then
-        if [ $ARCH == 32 ] ; then
-            mkdir -p $1/lib/$TYPE/Win32
-            # copy .lib and .dll artifacts
-            cp -v build_vs_32/lib/Release/*.lib $1/lib/$TYPE/Win32
-            # copy header files
-            cp -v -r build_vs_32/include/* $1/include
-        elif [ $ARCH == 64 ] ; then
-            mkdir -p $1/lib/$TYPE/x64
-            # copy .lib and .dll artifacts
-            cp -vr build_vs_64/lib/Release $1/lib/$TYPE/x64
-            cp -vr build_vs_64/lib/Debug $1/lib/$TYPE/x64
-            # copy header files
-            cp -v -r build_vs_64/include/* $1/include
-        fi
+    if [ "$TYPE" == "vs" ] ; then            
+        cp -v -r build_${TYPE}_${ARCH}/include/* $1/include
+        mkdir -p $1/lib/$TYPE/$PLATFORM/
+        cp -v "build_${TYPE}_${ARCH}/lib/Release/assimp-vc${VC_VERSION}-mt.lib" $1/lib/$TYPE/$PLATFORM/libassimp.lib  
     elif [ "$TYPE" == "osx" ] ; then
         cp -Rv lib/libassimp.a $1/lib/$TYPE/assimp.a
     elif [[ "$TYPE" == "ios" || "$TYPE" == "tvos" ]] ; then
@@ -247,15 +323,17 @@ function copy() {
         cp -Rv include/* $1/include
     elif [ "$TYPE" == "android" ]; then
         mkdir -p $1/lib/$TYPE/$ABI/
-        cp -Rv build_android/include/* $1/include
-        cp -Rv build_android/lib/libassimp.a $1/lib/$TYPE/$ABI/libassimp.a
+        cp -Rv build_${TYPE}_${ABI}/include/* $1/include
+        cp -Rv build_${TYPE}_${ABI}/lib/libassimp.a $1/lib/$TYPE/$ABI/libassimp.a
     elif [ "$TYPE" == "emscripten" ]; then
         cp -Rv build_emscripten/include/* $1/include
-        cp -Rv build_emscripten/lib/libassimp.a $1/lib/$TYPE/libassimp.a
+        cp -v "build_${TYPE}/lib/libassimp.a" $1/lib/$TYPE/libassimp.a
     fi
 
     # copy license files
-    rm -rf $1/license # remove any older files if exists
+    if [ -d "$1/license" ]; then
+        rm -rf $1/license
+    fi
     mkdir -p $1/license
     cp -v LICENSE $1/license/
 }
@@ -264,20 +342,35 @@ function copy() {
 function clean() {
 
     if [ "$TYPE" == "vs" ] ; then
-        if [ $ARCH == 32 ] ; then
-            vs-clean "build_vs_32/Assimp.sln";
-        elif [ $ARCH == 64 ] ; then
-            vs-clean "build_vs_64/Assimp.sln";
-        fi
+        rm -f build_${TYPE}_${ARCH}
         rm -f CMakeCache.txt
         echo "Assimp VS | $TYPE | $ARCH cleaned"
 
     elif [ "$TYPE" == "android" ] ; then
-        echoWarning "TODO: clean android"
+        if [ -d "build" ]; then
+            cd  "build_${TYPE}_${ABI}"
+            make clean
+            cd ..
+        fi
+        rm -f CMakeCache.txt  2> /dev/null
 
     else
         make clean
         make rebuild_cache
-        rm -f CMakeCache.txt
+        rm -f CMakeCache.txt 2> /dev/null
+    fi
+}
+
+function save() {
+    . "$SAVE_SCRIPT" 
+    savestatus ${TYPE} "assimp" ${ARCH} ${VER} true "${SAVE_FILE}"
+}
+
+function load() {
+    . "$LOAD_SCRIPT"
+    if loadsave ${TYPE} "assimp" ${ARCH} ${VER} "${SAVE_FILE}"; then
+      return 0;
+    else
+      return 1;
     fi
 }
