+++ apothecary/formulas/glm/glm.sh
@@ -9,7 +9,7 @@ FORMULA_TYPES=( "osx" "linux" "linux64" "linuxarmv6l" "linuxarmv7l" "linuxaarch6
 
 # tools for git use
 GIT_URL=https://github.com/g-truc/glm
-#GIT_TAG=0.9.9.7
+#GIT_TAG=0.9.9.8
 GIT_TAG=master
 
 # download the source code and unpack it into LIB_NAME
@@ -35,9 +35,11 @@ function copy() {
 	cp -rv glm $1/include
 
 	# copy license file
-	rm -rf $1/license # remove any older files if exists
+	if [ -d "$1/license" ]; then
+        rm -rf $1/license
+    fi
 	mkdir -p $1/license
-	cp -v readme.md $1/license/
+	cp -v copying.txt $1/license/license.txt
 }
 
 # executed inside the lib src dir
